Move sources into lib, compile to dist
