+++ lib/bundleEPUB.js
