+++ lib/checkEPUB.js
