+++ lib/manifest.js
