+++ lib/renderEPUB.js
