+++ tsconfig.json
@@ -1,26 +1,23 @@
 {
-    // Change this to match your project
-    // "include": ["lib/**/*"],
     "files": [
-        "bundleEPUB.js", "checkEPUB.js", "Configuration.js",
-        "index.js", "manifest.js", "metadata.js", "opf.js",
-        "utils.js"
+        "lib/bundleEPUB.js", "lib/checkEPUB.js", "lib/cli.js",
+        "lib/Configuration.ts", "lib/index.js", "lib/manifest.js",
+        "lib/metadata.js", "lib/opf.js",
+        "lib/utils.js", "lib/watcher.mjs"
     ],
     "compilerOptions": {
-      // Tells TypeScript to read JS files, as
-      // normally they are ignored as source files
-      "allowJs": true,
-      // Generate d.ts files
-      "declaration": true,
-      // This compiler run should
-      // only output d.ts files
-      "emitDeclarationOnly": true,
-      // Types should go into this directory.
-      // Removing this would place the .d.ts files
-      // next to the .js files
-      "outDir": "types",
-      // go to js file when using IDE functions like
-      // "Go to Definition" in VSCode
-      "declarationMap": true
+        "lib": [ "es6", "es2021", "esnext" ],
+        "target": "es2021",
+        "module": "commonjs",
+        "moduleResolution": "Node",
+        "outDir": "./dist",
+        "allowJs": true,
+        "declaration": true,
+        "declarationMap": true,
+        "inlineSourceMap": true,
+        "inlineSources": true,
+        "emitDecoratorMetadata": true,
+        "experimentalDecorators": true,
+        "esModuleInterop": true
     }
 }
